Send reminder message to both #vc-events and optional second channel
The diff does not look nice, but there are two main changes:

For each event, instead of choosing between DEFAULT_SLACK_EVENT_CHANNEL
and event.eventSlackAnnouncementsChannelId (and preferring
event.eventSlackAnnouncementsChannelId over DEFAULT_SLACK_EVENT_CHANNEL)
we create a message for DEFAULT_SLACK_EVENT_CHANNEL and then if
event.eventSlackAnnouncementsChannelId is set, then we also create a
message for that channel.

We apply the same logic for the hourlyAdminMessage to generate a list of
channels it got posted to.